[procon test] Run test src/atcoder/abc045/a.py

+++ src/atcoder/abc045/a.py
@@ -2,7 +2,7 @@ a = int(input())
 b = int(input())
 h = int(input())
 
-ans = (a + b) * h / 2
+ans = int((a + b) * h / 2)
 
 print(ans)
 print()
